chore(structure): move code to src/{app,pipeline,tools}; data→raw/processed; init packages; update .gitignore
