+++ data/raw/articles/rektcapital_002.json
